Rename LibraryLoader to LibraryInjector.
